Index_creation/requirements.txt
# Azure SDK packages
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-core>=1.29.0

# OpenAI
openai>=1.12.0

# Image processing
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
